Mudando arquivo seed Usuários para TS e adicionando comando yarn seed
